refactoring of the CertificateForm component into smaller, more manageable components with a clear separation of concerns and add  react-hot-toast
