+++ client/src/index.css
@@ -18,19 +18,19 @@ body {
 
 /* Track (the background) */
 ::-webkit-scrollbar-track {
-  background: #f0f0f0;
-  border-radius: 10px;
+  background: #050029;
+   
 }
 
-/* Thumb (the draggable part) */
 ::-webkit-scrollbar-thumb {
   background-color: #8400ff;
-  
+  /* XROW */
+   box-shadow: 0px 0px 30px 30px #8400ff;
+
 }
 
-/* On hover */
 ::-webkit-scrollbar-thumb:hover {
-   background-color: #4a0090;
+  background-color: #7200dc; 
 }
