+++ package.json
@@ -12,6 +12,7 @@
     "@openzeppelin/contracts": "^5.2.0",
     "dotenv": "^16.5.0",
     "ipfs-http-client": "^60.0.1",
+    "react-hot-toast": "^2.5.2",
     "web3": "^4.5.0"
   },
   "devDependencies": {
